Go beyond the teacher's code and improve the program.
Edit the checkout.feature file to remove the unused "Name" from the Then step, and replace it with a "Quantity".
Add a column to the Examples table in checkout.feature so that the number of items can be passed as a parameter.
Edit CheckoutPage and CheckoutPageStepDefinition to make use of the new "Quantity" value and actually verify something.